make main function more slim
